Tentatively moved uuid type and implemented blob-based variant for sqlite
